Refactor code into a single js file
