+++ decidim-comments/app/assets/config/decidim_comments_manifest.js
@@ -1,1 +1,1 @@
-//= link decidim/comments/comments.js
+//= link decidim/meetings/map.js
