+++ decidim-meetings/app/assets/config/decidim_meetings_manifest.js
@@ -1,0 +1,2 @@
+//= link decidim/meetings/map.js
+//= link decidim/meetings/map.css
