+++ decidim-meetings/app/assets/javascripts/decidim/meetings/map.js.es6.erb
@@ -1,0 +1,55 @@
+// = require leaflet
+// = require leaflet-tilelayer-here
+// = require leaflet-svg-icon
+// = require jquery-tmpl
+
+/* globals L */
+
+$(document).on("turbolinks:load", () => {
+  const mapId = 'decidim-meetings-map';
+  const $map = $(`#${mapId}`);
+
+  if ($map.length > 0) {
+    const popupTemplateId = 'meeting-popup';
+    const meetingsData = $map.data('meetings');
+    const bounds = new L.LatLngBounds(meetingsData.map((meeting) => [meeting.latitude, meeting.longitude]));
+    const map = L.map(mapId);
+
+    $.template('meeting-popup', $(`#${popupTemplateId}`).html());
+
+    L.DivIcon.SVGIcon.DecidimIcon = L.DivIcon.SVGIcon.extend({
+      options: {
+        fillColor: '#ef604d',
+        opacity: 0
+      },
+      _createPathDescription: function() {
+        return 'M18,6.78a7.79,7.79,0,0,0-7.79,7.79c0,7.5,6.82,13.74,7.11,14a1,1,0,0,0,1.35,0c0.29-.29,7.11-6.54,7.11-14A7.79,7.79,0,0,0,18,6.78Zm0,11.61A3.39,3.39,0,1,1,21.39,15,3.39,3.39,0,0,1,18,18.39Z';
+      },
+      _createCircle: function() {
+        return ""
+      }
+    });
+
+    L.tileLayer.here({
+      appId: '<%= Rails.application.secrets.dig(:geocoder, "api_key").try(&:first) %>',
+      appCode: '<%= Rails.application.secrets.dig(:geocoder, "api_key").try(&:last) %>'
+    }).addTo(map);
+
+    meetingsData.forEach((meeting) => {
+      let marker = L.marker([meeting.latitude, meeting.longitude], { icon: new L.DivIcon.SVGIcon.DecidimIcon() }).addTo(map);
+      let node = document.createElement('div');
+
+      $.tmpl('meeting-popup', meeting).appendTo(node);
+
+      marker.bindPopup(node, {
+        maxwidth: 640,
+        minWidth: 500,
+        keepInView: true,
+        className: 'map-info'
+      }).openPopup();
+    });
+
+    map.fitBounds(bounds);
+    map.zoomOut();
+  }
+});
