+++ decidim-meetings/app/assets/stylesheets/decidim/meetings/map.css
@@ -1,0 +1,1 @@
+/* = require leaflet */
