+++ decidim-meetings/decidim-meetings.gemspec
@@ -20,6 +20,7 @@ Gem::Specification.new do |s|
   s.add_dependency "kaminari", "~> 1.0.1"
   s.add_dependency "geocoder", "~> 1.4.1"
   s.add_dependency "httparty", "~> 0.14.0"
+  s.add_dependency "jquery-tmpl-rails", "~> 1.1.0"
 
   s.add_development_dependency "decidim-dev", Decidim.version
   s.add_development_dependency "decidim-proposals", Decidim.version
